EX-128 - Created an extensions system for our emulation platform EX-98 - Created our initial version of a generic Mega Drive ROM loader using the extensions system EX-106 - Implemented correct support for HV counter latching in VDP core. This fixes the Sunset Riders intro screen EX-104 - Added support for creating and removing breakpoints from the disassemly window. Also implemented support for enabling and disabling breakpoints, which can also be done from the disassembly window. EX-102 - Added location masks to breakpoints and watchpoints EX-89 - Fixed the 32-bit build target EX-103 - Added the display of module names to the loading/unloading progress windows EX-129 - Fixed some system deadlock cases EX-130 - Added the correct power-on initialization of VRAM, CRAM, and VSRAM buffers in VDP core. Also implemented correct power-on initialization of HV counter in VDP core. EX-131 - Implemented correct power-on initialization for M68000 registers EX-132 - Fixed errors in M68000 MOVE and MOVEA instructions that were causing incorrect execution times to be calculated EX-133 - Fixed default encoding of XML data so that ASCII XML files can now be loaded EX-134 - Fixed an error which prevented the event logger showing all log messages EX-135 - Made the disassembly window show some opcodes above the currently selected opcode by default EX-136 - Added support for scrolling using keys and the mouse wheel in the gridlist control EX-137 - Some improvements and fixes to the VDP Image window EX-138 - Added initial support for apparent power-on initialization delay in M68000 core. Further testing and improvements required.

+++ Devices/M68000/ExceptionsView.cpp
@@ -213,7 +213,7 @@ INT_PTR M68000::ExceptionsView::msgWM_COMMAND(HWND hwnd, WPARAM wparam, LPARAM l
 				device->exceptionListEmpty = device->exceptionList.empty();
 				exceptionListCopy = device->exceptionList;
 				SendMessage(GetDlgItem(hwnd, IDC_M68000_EXCEPTIONS_LIST), LB_DELETESTRING, exceptionListIndex, NULL);
-				if(exceptionListIndex >= exceptionListCopy.size())
+				if(exceptionListIndex >= (int)exceptionListCopy.size())
 				{
 					exceptionListIndex -= 1;
 				}
